cli: add CLIPlugin to specify resources via shape or json
Problem: currently, users have no way on the CLI of utilizing
the experimental parser in RFC 46 for specifying a job shape
or a resource set they'd like to use via JSON. This causes
users to defer to cumbersome piping and JSON mangling when
they want to submit jobs of arbitrary resource shapes on the
command line.

Add a /plugins subdirectory to src/bindings/python/flux/cli for
optional plugins that enable experimental functionality. Add a
plugin for specifying `--r-json` or `--r-shape` for specifying
resources on the command line beyond the standard ones supported.